refactor utils export
https://coveord.atlassian.net/browse/KIT-4258
