+++ packages/atomic/src/utils/item-section-utils.ts
@@ -1,4 +1,4 @@
-import {containsVisualElement} from './stencil-utils';
+import {containsVisualElement} from './utils';
 
 export function hideEmptySection(element: HTMLElement) {
   element.style.display = containsVisualElement(element) ? '' : 'none';
